backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
python-multipart
google-genai
google-cloud-bigquery
openai
pymupdf
